Move GlobalExceptionHandlerAdapter to core package
