+++ src/main/java/com/freya02/botcommands/api/core/GlobalExceptionHandlerAdapter.java
@@ -1,6 +1,6 @@
-package com.freya02.botcommands.api;
+package com.freya02.botcommands.api.core;
 
-import com.freya02.botcommands.api.core.GlobalExceptionHandler;
+import com.freya02.botcommands.api.BContext;
 import net.dv8tion.jda.api.events.Event;
 import net.dv8tion.jda.api.events.interaction.ModalInteractionEvent;
 import net.dv8tion.jda.api.events.interaction.command.MessageContextInteractionEvent;
