+++ src/test/kotlin/com/freya02/botcommands/test_kt/MyGlobalExceptionHandler.kt
@@ -1,7 +1,7 @@
 package com.freya02.botcommands.test_kt
 
 import com.freya02.botcommands.api.BContext
-import com.freya02.botcommands.api.GlobalExceptionHandlerAdapter
+import com.freya02.botcommands.api.core.GlobalExceptionHandlerAdapter
 import com.freya02.botcommands.api.core.service.annotations.BService
 import mu.KotlinLogging
 import net.dv8tion.jda.api.events.Event
